Auth moved to the separate package
