+++ go.mod
@@ -6,8 +6,11 @@ require (
 	github.com/gbrlsnchs/jwt/v3 v3.0.0-beta.0
 	github.com/gorilla/mux v1.7.1
 	github.com/gorilla/websocket v1.4.0
+	github.com/kr/pretty v0.1.0 // indirect
 	github.com/lib/pq v1.0.0
 	github.com/manveru/faker v0.0.0-20171103152722-9fbc68a78c4d
+	github.com/manveru/gobdd v0.0.0-20131210092515-f1a17fdd710b // indirect
 	github.com/satori/go.uuid v1.2.0
 	github.com/xlab/closer v0.0.0-20190328110542-03326addb7c2
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
